app/requirements.txt
fastapi
mysql-connector-python
python-dotenv
kafka-python
uvicorn